Loading and Restoring
I made the loading screen look better, then added the ability for idle shop items to spawn in from the save code and the upgrades. At some point I caused a bug with the CPS counter, but that only happens if you restore from a save code with an upgrade and no idle items, which is unlikely unless you are hacking.